+++ Src/Engine/Server.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "Tools.hpp"
 #include "Network.hpp"
 #include <chrono>
 
@@ -14,7 +15,29 @@ struct client_t {
     std::string name = "";
     bool pending_disconnect = false;
     sc::time_point disconnect_at { };
+    void* data = nullptr;
+};
+
+enum {
+    DISC_OK = 0,
+    DISC_EOP,
+    DISC_MSGERR,
+    DISC_SERVER_FULL,
+    DISC_TIMEOUT,
+    DISC_UNKNOWN,
+    NUMDISCS
+};
+
+constexpr const char* disconnect_reason[NUMDISCS] = {
+    "",
+    "end of packet",
+    "server is full",
+    "message error",
+    "connection timed out"
+    "unknown reason"
 };
 
 extern std::vector<client_t> clients;
-extern void handle_packets(ENetPacket* packet, ENetPeer* peer);
+extern void do_disconnect(client_iterator&);
+extern result<void, int> handle_packets(packetstream packet, ENetPeer* peer);
+