+++ Src/Game/Game.hpp
@@ -110,8 +110,8 @@ enum {
     N_UPDATE_ENTITIES,
     N_PING,
     N_PONG,
+    N_ROUNDINFO,
     N_STARTROUND,
-    N_SET_AUTOSTART,
     N_SET_SPEED,
     N_PAUSE,
     N_RESUME,
@@ -123,29 +123,30 @@ enum {
     NUMMSG
 };
 
-constexpr uint32_t msgsizes[NUMMSG] = { // Message sizes in bytes, -1 for unchecked/variable sizes
-    /* N_CONNECT: */               -1,
-    /* N_TEXT: */                  -1,
-    /* N_PLACETOWER: */            20,
-    /* N_UPGRADETOWER: */           5,
-    /* N_UPDATETARGETING: */        5,
-    /* N_SELLTOWER: */              4,
-    /* N_UPDATE_CASH: */            8,
-    /* N_UPDATE_LIVES: */           8,
-    /* N_REQUEST_UPDATE: */         0,
-    /* N_UPDATE_ENTITIES: */       -1,
-    /* N_PING: */                   4,
-    /* N_PONG: */                   4,
-    /* N_STARTROUND: */             0,
-    /* N_SET_AUTOSTART: */          1,
-    /* N_SET_SPEED: */              1,
-    /* N_PAUSE: */                  0,
-    /* N_RESUME: */                 0,
-    /* N_ROUNDOVER: */              0,
-    /* N_GAMEOVER: */               0, // This will probably send some stuff later like what enemy was leaked and some kind of score
-    /* N_RESTART: */                0,
-    /* N_CONTINUE: */               0,
-    /* N_DISCONNECT: */             0
+// Message sizes in bytes, -1 for unchecked/variable sizes. First value is for the server, second client.
+constexpr uint32_t msgsizes[NUMMSG][2] = {
+    /* N_CONNECT: */               { -1, -1 },
+    /* N_TEXT: */                  { -1, -1 },
+    /* N_PLACETOWER: */            { 20, 20 },
+    /* N_UPGRADETOWER: */          {  5,  5 },
+    /* N_UPDATETARGETING: */       {  5,  5 },
+    /* N_SELLTOWER: */             {  4,  4 },
+    /* N_UPDATE_CASH: */           {  8, 12 },
+    /* N_UPDATE_LIVES: */          {  8,  8 },
+    /* N_REQUEST_UPDATE: */        {  0,  0 },
+    /* N_UPDATE_ENTITIES: */       { -1, -1 },
+    /* N_PING: */                  {  4,  4 },
+    /* N_PONG: */                  {  4,  4 },
+    /* N_ROUNDINFO: */             {  4,  4 },
+    /* N_STARTROUND: */            {  0,  0 },
+    /* N_SET_SPEED: */             {  1,  1 },
+    /* N_PAUSE: */                 {  0,  4 },
+    /* N_RESUME: */                {  0,  4 },
+    /* N_ROUNDOVER: */             {  0,  0 },
+    /* N_GAMEOVER: */              {  0,  0 }, // This will probably send some stuff later like what enemy was leaked and some kind of score
+    /* N_RESTART: */               {  0,  4 },
+    /* N_CONTINUE: */              {  0,  0 },
+    /* N_DISCONNECT: */            {  0,  0 }
 };
 
 static const struct enemy_t {
@@ -1972,3 +1973,14 @@ struct game_state {
     std::chrono::system_clock::time_point last_tick;
     std::chrono::system_clock::time_point last_spawned_tick;
 };
+
+struct gamestate {
+    double lives      = 0.0;
+    bool running      = false;
+    bool paused       = false;
+    bool fast_forward = false;
+    size_t cur_round  = 0;
+    size_t last_round = -1;
+};
+
+extern gamestate gs;